add instrument downstream task

+++ s3prl/s3prl/run_instrument.py
@@ -1,23 +1,38 @@
 import os
 
+not_download, not_preprocess = False, False
+train = True
 if __name__ == "__main__":
-    os.chdir("data")
-    os.system(
-        "wget https://zenodo.org/record/1290750/files/IRMAS-TestingData-Part1.zip?download=1 -O IRMAS-TestingData-Part1.zip"
-    )
-    os.system(
-        "wget https://zenodo.org/record/1290750/files/IRMAS-TestingData-Part2.zip?download=1 -O IRMAS-TestingData-Part2.zip"
-    )
-    os.system(
-        "wget https://zenodo.org/record/1290750/files/IRMAS-TestingData-Part3.zip?download=1 -O IRMAS-TestingData-Part3.zip"
-    )
-    os.system(
-        "wget https://zenodo.org/record/1290750/files/IRMAS-TrainingData.zip?download=1 -O IRMAS-TrainingData.zip"
-    )
+    # Download IRMAS dataset
+    if not_download:
+        os.chdir("data")
+        os.system(
+            "wget https://zenodo.org/record/1290750/files/IRMAS-TestingData-Part1.zip?download=1 -O IRMAS-TestingData-Part1.zip"
+        )
+        os.system(
+            "wget https://zenodo.org/record/1290750/files/IRMAS-TestingData-Part2.zip?download=1 -O IRMAS-TestingData-Part2.zip"
+        )
+        os.system(
+            "wget https://zenodo.org/record/1290750/files/IRMAS-TestingData-Part3.zip?download=1 -O IRMAS-TestingData-Part3.zip"
+        )
+        os.system(
+            "wget https://zenodo.org/record/1290750/files/IRMAS-TrainingData.zip?download=1 -O IRMAS-TrainingData.zip"
+        )
 
-    os.system("unzip IRMAS-TestingData-Part1.zip && rm IRMAS-TestingData-Part1.zip")
-    os.system("unzip IRMAS-TestingData-Part2.zip && rm IRMAS-TestingData-Part2.zip")
-    os.system("unzip IRMAS-TestingData-Part3.zip && rm IRMAS-TestingData-Part3.zip")
-    os.system("unzip IRMAS-TrainingData.zip && rm IRMAS-TrainingData.zip")
+        os.system("unzip IRMAS-TestingData-Part1.zip && rm IRMAS-TestingData-Part1.zip")
+        os.system("unzip IRMAS-TestingData-Part2.zip && rm IRMAS-TestingData-Part2.zip")
+        os.system("unzip IRMAS-TestingData-Part3.zip && rm IRMAS-TestingData-Part3.zip")
+        os.system("unzip IRMAS-TrainingData.zip && rm IRMAS-TrainingData.zip")
+
+    # process IRMAS dataset
+    if not_preprocess:
+        os.system("python downstream/instrument_identification/dataset/prepare_data.py")
+
+    # train
+    if train:
+        os.system(
+            "python run_downstream.py -m train -n hubert--instrument_identification \
+                -u hubert -d instrument_identification -cd"
+        )
 
     os.system("echo 'Done!' ")
